% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_avbiomsp}
\alias{atlantis_avbiomsp}
\title{Test if the functional group biomass collapsed.}
\usage{
atlantis_avbiomsp(func_grp, path, exe_name, run_time, batch_file = NULL)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{run_time}{The total duration of the simulation in days
(\code{burnin_time} + 10950).}

\item{batch_file}{The name of the batch/shell file you are using to run your
model. If not provided, the function will search for the unique batch file
in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{biom_sp} The average biomass of the 5 last years of simulation.
  Return 0 if \code{biom_sp} < 0.1.
}
\description{
It opens the output file \code{BiomIndx.txt} and calculates the
  average biomass of the selected functional group \code{func_grp} over the 5
  last years of simulation. If the average biomass is smaller than 0.1 metric
  tons, the stock is considered collapsed and it returns 0.
}
\examples{
atlantis_avbiomsp("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                  "atlantismain", 18250, "runAtlantis.bat")
atlantis_avbiomsp("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                  "atlantisNew", 18250, "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection, and
  \code{\link[utils]{read.table}} to open table stored in text files (.txt,
  .csv ...).
}
