% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_checkf}
\alias{atlantis_checkf}
\title{Test if model uses fishing mortality.}
\usage{
atlantis_checkf(path, harvest_filename, fishing_para)
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{harvest_filename}{The name of the harvest parameters file with its
extension (ex: file.prm).}

\item{fishing_para}{A dataframe containing data from the fishing fleet csv
file plus an extra column to indicate if the fleet is active (1) or not (0)
named \code{active_flt} (output from the function
\code{\link{atlantis_fleetopen}}).}
}
\value{
\code{fishing_para} The input dataframe \code{fishing_para} with an
  extra column named \code{effortmodel} containing the effortmodel option
  used for each fleet.
}
\description{
Test if the active fleet uses the fishing mortality parameter in
  Atlantis. Check in the harvest parameters file if the parameter
  \code{XXX_effortmodel} is equal to 0 (Fishing mortality option) for each
  active fleet.
}
\examples{
atlantis_checkf("C:/Atlantis/AtlantisEEC/AtlantisEECF_v3", "AEEC_harvest.prm",
                fishing_para)
atlantis_checkf("/home/Atlantis/AtlantisEEC/AtlantisEECF_v3", "AEEC_harvest.prm",
                fishing_para)
}
\seealso{
\code{\link{atlantis_openfile}} to open a parameters file and select
  a parameter, \code{\link[stringr]{str_split_fixed}} to split a chain of
  characters, and \code{\link[qdapRegex]{ex_number}} to extract number from a
  chain of characters.
}
