% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_checkmodule}
\alias{atlantis_checkmodule}
\title{Test if Atlantis modules are on.}
\usage{
atlantis_checkmodule(model_path, exe_name, batch_file = NULL)
}
\arguments{
\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{test} A binary variable (1) all the modules are on (0) at least
  one module is off.
}
\description{
Test if at least physics, biology and fishery modules are on in
  the calibrated model. It looks inside the run parameters file and check if
  \code{flag_fisheries_on} = 1, \code{flag_skip_biol} = 0, and
  \code{flag_skip_phys} = 0.
}
\examples{
atlantis_checkmodule("C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                     "atlantismain", "runAtlantis.bat")
atlantis_checkmodule("/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                     "atlantisNew", "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection, and
  \code{\link{atlantis_openfile}} to open a parameters file and select a
  parameter.
}
