% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_diet.R
\name{atlantis_dietextract}
\alias{atlantis_dietextract}
\title{Extract availability matrix from biological parameters file}
\usage{
atlantis_dietextract(path, exe_name, save.out = F, output_path = NULL)
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{save.out}{Flag that indicate if you want to save the output or not.
If \code{TRUE}, \code{output_path} need to be provided.
(Default value is \code{FALSE}).}

\item{output_path}{The directory were you want to save the csv file
(Default value is \code{NULL}).}
}
\value{
\code{availability} Dataframe containing input availability matrix
from the biological parameters file.
}
\description{
Extract diet from Biological parameters file
and save it if wanted into a readable csv file.
}
\examples{
  atlantis_dietextract("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantismain")
  atlantis_dietextract("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantisNew")
  atlantis_dietextract("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantismain", save.out = T, output_path =
                    "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
  atlantis_dietextract("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantisNew", save.out = T, output_path =
                    "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link{atlantis_openfile}} to open a parameters file and select a parameter,
  \code{\link[stringr]{str_split_fixed}} to split a chain of characters,
  \code{\link[utils]{read.table}} to open table stored in text files (.txt, .csv ...).
}
