% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_core.R
\name{atlantis_fchange}
\alias{atlantis_fchange}
\title{Change F in Atlantis.}
\usage{
atlantis_fchange(func_grp, path, harvest_filename, f_prop, f_test,
  fishing_para)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{harvest_filename}{The name of the harvest parameters file with its
extension (ex: file.prm).}

\item{f_prop}{The proportion of the total fishing pressure applied per each
fleet. The order of the fleets is the same as the one used in Atlantis.}

\item{f_test}{The new value of total fishing mortality to be tested in days.}

\item{fishing_para}{A dataframe containing data from the fishing fleet csv
file plus two extra column, one to indicate if the fleet is active (1) or
not (0) named \code{active_flt} and one named \code{effortmodel} containing
the effortmodel option used for each fleet (output from the function
\code{\link{atlantis_checkf}}).}
}
\value{
Change value of F in the harvest parameters files.
}
\description{
It modifies the harvest parameters file by applying a new value
  of F for the functional group \code{func_grp}. It looks for \code{mFC_XXX}
  parameter and and change the value of the vector using the new F value
  \code{f_test} that need to be simulated and the distribution of the fishing
  pressure accross fleets \code{f_prop} from the calibrated model.
}
\examples{
atlantis_fchange("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "AEEC_harvest.prm", c(0.2, 0.45, 0.05, 0.12, 0.08, 0.1), 0.0002739726,
  fishing_para)
atlantis_fchange("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "AEEC_harvest.prm", c(0.2, 0.45, 0.05, 0.12, 0.08, 0.1), 0.0002739726,
  fishing_para)
}
\seealso{
\code{\link{atlantis_openfile}} to open a parameters file and select
  a parameter.
}
