% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_ini.R
\name{atlantis_fdistri}
\alias{atlantis_fdistri}
\title{Atlantis F distribution.}
\usage{
atlantis_fdistri(func_grp, model_path, harvest_filename, fishing_para,
  raw.distri = F)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{harvest_filename}{The name of the harvest parameters file with its
extension (ex: file.prm).}

\item{fishing_para}{A dataframe containing data from the fishing fleet csv
file plus two extra columns, one to indicate if the fleet is active (1) or
not (0) named \code{active_flt} and one named \code{effortmodel} containing
the effortmodel option used for each fleet (output from the function
\code{\link{atlantis_checkf}}).}

\item{raw.distri}{If TRUE return the raw value of fishing mortality (F) for
active fleets (non-active fleets are set to 0), else the distribution of F is
returned (vector of values between [0-1] and their sum is 1). \strong{Default:}
FALSE.}
}
\value{
\code{f_prop} A vector with the proportion of fishing pressure
  applied per each fleet. The order of the fleets is the same as the one used
  in Atlantis.
}
\description{
Extract the fishing mortality distribution accross fleets in
  Atlantis. It opens the harvest parameters file, select the \code{mFC_XXX}
  vector for the considered functional group \code{func_grp} and return the
  proportion of the total F allocated to each fleet. If the functional group
  \code{func_grp} is unfished in the calibrated model this function will
  attribute fishing pressure evenly accross active fleets.
}
\examples{
atlantis_fdistri("COD", "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
 "AEEC_harvest.prm", fishing_para)
atlantis_fdistri("COD", "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
 "AEEC_harvest.prm", fishing_para)
}
\seealso{
\code{\link{atlantis_openfile}} to open a parameters file and select
  a parameter.
}
