% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_fgrpimp}
\alias{atlantis_fgrpimp}
\title{Test if functional groups is fished and impacted.}
\usage{
atlantis_fgrpimp(func_grp, path, exe_name, batch_file = NULL)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{fgrpimp} A binary variable telling if the functional group is
  (1) fished and impacted or not (0).
}
\description{
Test if the functional group \code{func_grp} is fished and
  impacted in the calibrated model. It open the functional groups csv
  parameters file and look if the parameter \code{IsFished} and
  \code{IsImpacted} are on.
}
\examples{
atlantis_fgrpimp("COD", "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                 "atlantismain", "runAtlantis.bat")
atlantis_fgrpimp("COD", "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                 "atlantisNew", "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection, and
  \code{\link[utils]{read.table}} to open table stored in text files (.txt,
  .csv ...).
}
