% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_fgrpon}
\alias{atlantis_fgrpon}
\title{Test if functional groups is active.}
\usage{
atlantis_fgrpon(func_grp, path, exe_name)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}
}
\value{
\code{fgrpon} A binary variable telling if the functional group is
  (1) activated or not (0).
}
\description{
Test if the functional group \code{func_grp} is active in the
  calibrated model. It open the functional groups csv parameters file and
  look if the functional group is on.
}
\examples{
atlantis_fgrpon("COD", "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                "atlantismain")
atlantis_fgrpon("COD", "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                "atlantisNew")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection, and
  \code{\link[utils]{read.table}} to open table stored in text files (.txt,
  .csv ...).
}
