% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantis_fleetopen}
\alias{atlantis_fleetopen}
\title{Test if the fleets are active.}
\usage{
atlantis_fleetopen(path, exe_name, harvest_filename, run_time,
  batch_file = NULL)
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{harvest_filename}{The name of the harvest parameters file with its
extension (ex: file.prm).}

\item{run_time}{The total duration of the simulation in days
(\code{burnin_time} + 10950).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{fishing_para} A dataframe containing data from the fishing
  fleet csv file plus an extra column to indicate if the fleet is active (1)
  or not (0) named \code{active_flt}.
}
\description{
Test if at least one fleet is open in Atlantis. It opens the
  harvest parameters file to check the value of the the parameter
  \code{XXX_tStart} for each fleet. If the value of \code{XXX_tStart} is
  inferior to \code{run_time} the fleet is considered active. It opens as
  well the fishing fleet .csv file to obtain the characteristic of each fleet
  and add a column named \code{active_flt} to store the result of the test.
}
\examples{
atlantis_fleetopen("C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                   "atlantismain", "AEEC_harvest.prm", 18250, "runAtlantis.bat")
atlantis_fleetopen("/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                   "atlantisNew", "AEEC_harvest.prm", 18250, "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link{atlantis_openfile}} to open a parameter file and select a
  parameter, \code{\link[stringr]{str_split_fixed}} to split a chain of
  characters, \code{\link[qdapRegex]{ex_number}} to extract number from a
  chain of characters, and \code{\link[utils]{read.table}} to open table
  stored in text files (.txt, .csv ...).
}
