% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_fscenarios.R
\name{atlantis_fsimu}
\alias{atlantis_fsimu}
\title{Run set of fishing strategies scenarios.}
\usage{
atlantis_fsimu(
  func_grp,
  folder_path,
  model_path,
  exe_name,
  burnin_time,
  label,
  batch_file = NULL,
  fmult,
  fref = rep(1, length(fmult))
)
}
\arguments{
\item{func_grp}{The code of Atlantis functional groups for which F scenarios will
be run.}

\item{folder_path}{The working directory: where the F simulations will be
run and stored.}

\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew, atlantisMerged).}

\item{burnin_time}{The time in days needed for the model to pseudo-stabilize
before starting the additional 30 years of simulation.}

\item{label}{Name of the scenarios (used to set names of output folder and files).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}

\item{fmult}{Set of F multipliers, one value per functional groups in \code{func_grp}.
If wants to be used as set of fishing mortalities (in y-1), don't supply values for
\code{fref}.}

\item{fref}{Set of reference F (in y-1), one value per functional groups in
\code{func_grp}. The default value is set to 1.}
}
\value{
Atlantis outputs for every scenarios defined previously.
function(func_grp, folder_path, model_path, exe_name, burnin_time, label, batch_file = NULL, fmult, fref = rep(1,length(fmult)))
}
\description{
Run a predefine set of fishing scenarios on a several functional
groups using Fmsy multipliers or F per functional group. Unselected functional
groups remains at the same fishing level as in the initial model. F distribution
across fleet isn't changed.
}
\section{Warning}{

  At the moment, the package is not designed to run on OSX, it only
  works on Linux and Windows (To run on OSX these functions need to be
  modified: \code{\link{atlantisfmsy_simu}},
  \code{\link{atlantisfmsy_restart}}, \code{\link{atlantisfmsy_modelcopy}},
  \code{\link{atlantis_bachchange}}, and \code{\link{atlantis_paraselect}}).
  As well if you are not providing the name of your bach/shell file, this
  package required that only one batch/shell file is in the model directory
  (if more than one batch/shell files are present please remove them, leave
  only the one used to run the calibrated model).
}

\examples{
atlantisf_simu(c("COD","LBT"), "C:/Atlantis/AtlantisEEC/Fmsy",
                  "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                  "atlantismain", 7300, "HTL", "runAtlantis.bat")
atlantisf_simu(c("COD","LBT"), "/home/Atlantis/AtlantisEEC/Fmsy",
                  "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                  "atlantisNew", 7300, "HTL", "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_checkmodule}} to check if all the modules are
  on (physics, biology, and fishery), \code{\link{atlantis_fgrpon}} to check
  if the functional group is active, \code{\link{atlantis_fgrpimp}} to check
  if the functional group is fished and impacted,
  \code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link{atlantis_fleetopen}} to check if fleets are active,
  \code{\link{atlantis_checkf}} to check the formulation of the fishing
  pressure used (i.e. F, effort, dynamics model...),
  \code{\link{atlantis_fdistri}} to extract the distribution of fishing
  pressure amongst fleets, \code{\link{atlantisf_modelcopy}} to create the simulation
  directory and copy the calibrated model inside it, \code{\link{atlantis_runtime}}
  to modify the runtime in the run parameters file, \code{\link{atlantis_wsummary}}
  to modify periodicity of summary outputs, \code{\link{atlantis_fchange}}
  to change the value of F, \code{\link{atlantis_bachchange}} to change the
  name of output files, \code{\link[stringr]{str_split_fixed}} to split a chain
  of characters, \code{\link[qdapRegex]{ex_number}} to extract number from a chain of
  characters, \code{\link[ncdf4]{nc_open}} to open netCDF4 files,
  \code{\link[ncdf4]{ncvar_get}} to extract variables from netCDF4 files,
  \code{\link[ncdf4]{nc_close}} to close netCDF4 files, and
  \code{\link[utils]{read.table}} to open table stored in text files (.txt,
  .csv ...).
}
