% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_simurestart.R
\name{atlantis_lastsimu}
\alias{atlantis_lastsimu}
\title{Look for last F simulated.}
\usage{
atlantis_lastsimu(output_path, last_run = NULL)
}
\arguments{
\item{output_path}{The directory of the simulations outputs for the
functional group considered.}

\item{last_run}{The name of the output file of the last simulation run before
\code{\link{atlantisfmsy_simu}} stopped. The name of the output have to be
the same as the one in the batch/shell file used to run Atlantis. The
default value is NULL.}
}
\value{
A list containing a vector with the names of each files in
  \code{output_path} and the value of the last F simulated in y-1.
}
\description{
It looks for the last value of F simulated. It searches through
  all the output files generated by Atlantis in the \code{output_path}. And
  looks for the files created the most recently using \code{\link[base]{file.info}}.
}
\examples{
atlantis_lastsimu("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
atlantis_lastsimu("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
}
\seealso{
\code{\link[base]{file.info}} to recover files description.
}
