% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_fileselect.R
\name{atlantis_openfile}
\alias{atlantis_openfile}
\title{Open parameters file and look for a specific parameter.}
\usage{
atlantis_openfile(path, file_names, para_name)
}
\arguments{
\item{path}{The directory of the parameters file.}

\item{file_names}{The name of the parameters file with its extension (ex:
file.txt or file.prm).}

\item{para_name}{The name of the parameter the function will look for.}
}
\value{
a list containing: the open parameters file and the line ID (location
  of the parameter in the file).
}
\description{
It opens a parameters file, removes comment lines and finds the
  line ID of the parameter looked for.
}
\examples{
atlantis_openfile("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD", "AEEC35_run.prm", "tstop")
atlantis_openfile("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD", "AEEC35_run.prm", "tstop")
}
