% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_fileselect.R
\name{atlantis_paraselect}
\alias{atlantis_paraselect}
\title{Select Atlantis parameters file.}
\usage{
atlantis_paraselect(path, exe_name, index, batch_file = NULL)
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{index}{The index used by Atlantis to determine the type of the
different parameters files in the batch/shell file (ex: -i initial
condition file; -o output files; -r run file; -f forcing file; -p physics
file; -b biology file; -s functional groups csv file; -h fishing file; -q
fisheries csv file...).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{infilename} The name of the parameters file selected.
}
\description{
It looks for the names of the parameters file wanted within the
  batch or shell file used to run Atlantis.
}
\examples{
atlantis_paraselect("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantismain", "-h", "runAtlantis.bat")
atlantis_paraselect("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantisNew", "-h", "runAtlantis.sh")

}
