% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_ini.R
\name{atlantis_savestart}
\alias{atlantis_savestart}
\title{Change number of days to save from model outputs.}
\usage{
atlantis_savestart(path, exe_name, run_time, save_time, batch_file = NULL)
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{run_time}{The total duration of the simulation in days
(\code{burnin_time} + 10950).}

\item{save_time}{The number of days to save from the end of the simulation.}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
Modify run parameters file.
}
\description{
Change the output starting time in the run parameters file. It opens the run
  parameters file, select the \code{toutstart} parameter and replaces his value by:
  \code{runtime} = \code{burning_time} + 10950 days.
}
\examples{
atlantis_savestart("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantismain", 18250, 365, "runAtlantis.bat")
atlantis_savestart("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantisNew", 18250, 365, "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection, and
  \code{\link{atlantis_openfile}} to open parameter files and select
  parameters.
}
