% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_simudir.R
\name{atlantisf_modelcopy}
\alias{atlantisf_modelcopy}
\title{Copy the calibrated Atlantis model for F scenarios.}
\usage{
atlantisf_modelcopy(
  folder_path,
  model_path,
  exe_name,
  batch_file = NULL,
  folder_name
)
}
\arguments{
\item{folder_path}{The working directory: where the F simulations will be
run and stored.}

\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}

\item{folder_name}{The name of the simulation folder to contain a set of F scenarios.}
}
\value{
\code{simu_path} The simulation directory for the F scenarios considered.
}
\description{
Create the folder \code{folder_name} in the working directory
  \code{folder_path}. It copies/pastes inside it the files necessary to run
  Atlantis from the \code{model_path} directory.
}
\examples{
atlantisfmsy_modelcopy("C:/Atlantis/AtlantisEEC",
                       "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                       "atlantismain", "runAtlantis.bat", "HTL")
atlantisfmsy_modelcopy("/home/Atlantis/AtlantisEEC",
                       "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                       "atlantisNew", "runAtlantis.sh", "HTL")

}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link{atlantis_openfile}} to open a parameters file and select a
  parameter, \code{\link[utils]{tail}} to return the last part of an object
  and \code{\link[stringr]{str_split_fixed}} to split a chain of characters.
}
