% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_check.R
\name{atlantisfmsy_completion}
\alias{atlantisfmsy_completion}
\title{Test Fmsy simulation completion.}
\usage{
atlantisfmsy_completion(func_grp, output_path)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{output_path}{The directory of the output files.}
}
\value{
\code{end} A variable with three posibilities: (0) Fmsy isn't
  reached, (1) Fmsy is reached, and (2) no Fmsy have been found so fmax need
  to be increased.
}
\description{
It determines the F with the maximum cacth in the file
  \code{Fmsy_XXX.txt} and compares it to the two values of F planned for the
  next round of simulation stored in the file \code{Fnext_simu.txt}. If all
  the value are the same the Fmsy for the functional group \code{func_grp}
  has been found.
}
\examples{
atlantisfmsy_completion("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
atlantisfmsy_completion("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD/output")
}
\seealso{
\code{\link[utils]{read.table}} to open table stored in text files
  (.txt, .csv ...).
}
