% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_core.R
\name{atlantisfmsy_fmaxcatch}
\alias{atlantisfmsy_fmaxcatch}
\title{Find F with the maximum catch and estimate the next F values to simulate.}
\usage{
atlantisfmsy_fmaxcatch(func_grp, path, exe_name, run_time, fmax,
  batch_file = NULL)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{run_time}{The total duration of the simulation in days
(\code{burnin_time} + 10950).}

\item{fmax}{The maximal value of F the simulation is going to test in y-1.
\strong{WARNING:} only work if \code{fmax} < 10 y-1 (see
\code{\link{atlantis_bachchange}}).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
It writes on the drive, in the output folder, the file Fnext_simu.txt
  containing the functional group code \code{func_grp} (sp), the current F
  value with a maximum catch (F), and the two next F value to run (Ft1 and
  Ft2). It writes as well the file Fmsy_XXX.txt containing the functional
  group code \code{func_grp} (sp), the current F value with a maximum catch
  (F), and the yearly average catch over the last 5 years of simulation
  (yield).
}
\description{
It looks into the output files \code{...TOTCATCH.nc} and
  determine the value of F (which is coded in the output files name, see
  \code{\link{atlantis_bachchange}}) that result on the maximum catch for the
  functional group considered \code{func_grp}. Then it determines the two
  next values of F that need to be tested.
}
\examples{
atlantisfmsy_fmaxcatch("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantismain", 18250, 4, "runAtlantis.bat")
atlantisfmsy_fmaxcatch("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantisNew", 18250, 4, "runAtlantis.sh")

\dontrun{atlantisfmsy_fmaxcatch("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantismain", 18250, 10, "runAtlantis.bat")}
\dontrun{atlantisfmsy_fmaxcatch("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantisNew", 18250, 10, "runAtlantis.sh")}
}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link[ncdf4]{nc_open}} to open netCDF4 files,
  \code{\link[ncdf4]{ncvar_get}} to extract variables from netCDF4 files,
  \code{\link[ncdf4]{nc_close}} to close netCDF4 files,
  \code{\link[stats]{aggregate}} to aggregate data and
  \code{\link[utils]{write.table}} to write table on the drive.
}
