% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_core.R
\name{atlantisfmsy_inisimu}
\alias{atlantisfmsy_inisimu}
\title{F simulations up to the collapse of the stock.}
\usage{
atlantisfmsy_inisimu(
  func_grp,
  folder_path,
  model_path,
  exe_name,
  harvest_filename,
  run_time,
  f_prop,
  fmax,
  fmin,
  fishing_para,
  save_time = NULL,
  batch_file = NULL
)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{folder_path}{The working directory: where the Fmsy simulations will be
run and stored.}

\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{harvest_filename}{The name of the harvest parameters file with its
extension (ex: file.prm).}

\item{run_time}{The total duration of the simulation in days
(\code{burnin_time} + 10950).}

\item{f_prop}{The proportion of the total fishing pressure applied per each
fleet. The order of the fleets is the same as the one used in Atlantis.}

\item{fmax}{The maximal value of F the simulation is going to test in y-1.
\strong{WARNING:} only work if \code{fmax} < 10 y-1 (see
\code{\link{atlantis_bachchange}}).}

\item{fmin}{The minimum value of F to test in y-1 (only used when the
simulation are restarted after an interruption). The default value is 0.}

\item{fishing_para}{A dataframe containing data from the fishing fleet csv
file plus two extra columns, one to indicate if the fleet is active (1) or
not (0) named \code{active_flt} and one named \code{effortmodel} containing
the effortmodel option used for each fleet (output from the function
\code{\link{atlantis_checkf}}).}

\item{save_time}{The number of days to save from the end of the simulation.
\strong{Default:} NULL.}
}
\value{
\code{simu_path} The directory of the simulation folder, and run
  first set of simulations that are stored in the output folder.
}
\description{
In the directory \code{folder_path}, it creates the folder
  \code{AtlantisMSY} and the sub-folder named after the functional group
  \code{func_grp} where all simulations will take place. The calibrated model
  files from the directory \code{model_path} is copied into this newly
  created folder. Then a series of values of F between F = 0 to F =
  \code{fmax} with an incrementation of 0.4 y-1 are tested. When the stock
  collapse, the function stop and return the directory of the simulation
  folder.
}
\section{Warning}{

  Only work if fmax < 10 y-1, see \code{\link{atlantis_bachchange}}).
}

\examples{
atlantisfmsy_inisimu("COD", "C:/Atlantis/AtlantisEEC/",
 "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3", "atlantismain",
  "AEEC_harvest.prm", 18250, c(0.2, 0.45, 0.05, 0.12, 0.08, 0.1), 4, 0,
   fishing_para, "runAtlantis.bat")
atlantisfmsy_inisimu("COD", "/home/Atlantis/AtlantisEEC/",
 "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3", "atlantisNew",
  "AEEC_harvest.prm", 18250, c(0.2, 0.45, 0.05, 0.12, 0.08, 0.1), 4, 0,
   fishing_para, "runAtlantis.sh")

\dontrun{atlantisfmsy_inisimu("COD", "C:/Atlantis/AtlantisEEC/",
 "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3", "atlantismain",
  "AEEC_harvest.prm", 18250, c(0.2, 0.45, 0.05, 0.12, 0.08, 0.1), 10, 0,
   fishing_para)}
}
\seealso{
\code{\link{atlantisfmsy_modelcopy}} to create the simulation
  directory and copy the calibrated model inside it,
  \code{\link{atlantis_runtime}} to modify the runtime in the run parameters
  file, \code{\link{atlantis_wsummary}} to modify periodicity of summary
  outputs, \code{\link{atlantis_fchange}} to change the value of F in the
  harvest pramaters file, \code{\link{atlantis_bachchange}} to change output
  names in the batch/shell file, and \code{\link{atlantis_avbiomsp}} to
  estimate the functional group biomass.
}
