% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_simudir.R
\name{atlantisfmsy_modelcopy}
\alias{atlantisfmsy_modelcopy}
\title{Copy the calibrated Atlantis model.}
\usage{
atlantisfmsy_modelcopy(func_grp, folder_path, model_path, exe_name,
  batch_file = NULL, os = Sys.info()["sysname"])
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{folder_path}{The working directory: where the Fmsy simulations will be
run and stored.}

\item{model_path}{The directory of the calibrated model (containing all the
parameters files and one bach file. Forcing files can be stored in a direct
parent directory of model_path). \strong{WARNING:} Only working if the forcing
folder is in the main model directory \code{model_path} or if it is in the
direct parent directory. If not please either modify this package or modify
the path structure of your Atlantis input forcing parameters file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}

\item{os}{The operating system used (ex:"Windows" or "Linux"). \strong{WARNING:}
At the moment, the package is not designed to run on OSX (see
\code{\link{atlantis_bachchange}}, and \code{\link{atlantis_paraselect}}).}
}
\value{
\code{simu_path} The simulation directory for the functional group
  \code{func_grp} considered.
}
\description{
Create the folder AtlantisMSY in the working directory
  \code{folder_path} and a sub-folder named after the functional group code
  \code{func_grp} for the set of simulation. It copies/pastes inside it the
  files necessary to run Atlantis from the \code{model_path} directory.
}
\examples{
atlantisfmsy_modelcopy("COD", "C:/Atlantis/AtlantisEEC",
                       "C:/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                       "atlantismain", "runAtlantis.bat", "Windows")
atlantisfmsy_modelcopy("COD", "/home/Atlantis/AtlantisEEC",
                       "/home/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                       "atlantisNew", "runAtlantis.sh", "Linux")

\dontrun{atlantisfmsy_modelcopy("COD", "/Atlantis/AtlantisEEC",
                                "/Atlantis/AtlantisEEC/AtlantisEECF_v3",
                                "atlantismain", "Darwin") # for OSX.}
}
\seealso{
\code{\link{atlantis_paraselect}} for parameters file selection,
  \code{\link{atlantis_openfile}} to open a parameters file and select a
  parameter, \code{\link[utils]{tail}} to return the last part of an object
  and \code{\link[stringr]{str_split_fixed}} to split a chain of characters.
}
