% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_simudir.R
\name{pathconvert}
\alias{pathconvert}
\title{Convert paths into shortform on Windows to deals with space in path}
\usage{
pathconvert(path)
}
\arguments{
\item{path}{path to convert}
}
\value{
\code{path} converted path.
}
\description{
Convert the path \code{path} into the shortform to deals with space
  issues in Windows OS path. First the function check the OS to see if it's
  Windows and then it check if the path isn't NULL.
}
\note{
This is only working if the path already exist on your machine. If not, will
return path with double backslashes instead of slash but keep the space in the
path.
}
\examples{
pathconvert("C:/Atlantis/AtlantisEEC")

\dontrun{pathconvert("/home/Atlantis/AtlantisEEC") # for UNIX}

}
\seealso{
\code{\link[utils]{shortPathName}} to Convert file paths to the short
form.
}
