% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_core.R
\name{atlantis_bachchange}
\alias{atlantis_bachchange}
\title{Change bach/shell Atlantis file.}
\usage{
atlantis_bachchange(func_grp, path, exe_name, f_test, batch_file = NULL)
}
\arguments{
\item{func_grp}{The code of the Atlantis functional group for which Fmsy will
be estimated.}

\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{f_test}{The new value of total fishing mortality to be tested in days.}

\item{batch_file}{The name of the batch/shell file with extension you are using
to run your model. If not provided, the function will search for the unique
batch file in your \code{folder_path}. \strong{Default:} NULL.}
}
\value{
\code{f_test} The new value of total fishing mortality to be tested
  in days, and modify batch/shell Atlantis run file.
}
\description{
It modifies the name of the output files for the future run in
  the batch/shell file used to run Atlantis. The name is coded with the code
  of the functional group selected \code{func_grp} and the value of F that
  will be tested \code{f_test} (example: For \code{func_grp} = "COD" with a
  \code{f_test} = 0.0002739726 (0.1 y-1) the output file prefix will be
  MSY_COD_F010).
}
\section{Warning}{

  Only work if \code{fmax} < 10 y-1.
}

\examples{
atlantis_bachchange("COD", "C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantismain", 0.0002739726, "runAtlantis.bat")
atlantis_bachchange("COD", "/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
 "atlantisNew", 0.0002739726, "runAtlantis.sh")

}
\seealso{
\code{\link{atlantis_openfile}} to open a parameters file and select
  a parameter.
}
