% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantisfmsy_fileselect.R
\name{atlantis_paraselect}
\alias{atlantis_paraselect}
\title{Select Atlantis parameters file.}
\usage{
atlantis_paraselect(path, exe_name, index, os = Sys.info()["sysname"])
}
\arguments{
\item{path}{The directory of the batch or shell file.}

\item{exe_name}{The name of the atlantis executable you used (ex:
atlantismain, atlantisNew).}

\item{index}{The index used by Atlantis to determine the type of the
different parameters files in the batch/shell file (ex: -i initial
condition file; -o output files; -r run file; -f forcing file; -p physics
file; -b biology file; -s functional groups csv file; -h fishing file; -q
fisheries csv file...).}

\item{os}{The operating system used (ex:"Windows" or "Linux"). \strong{WARNING:}
At the moment, the package is not designed to run on OSX (see
\code{\link{atlantisfmsy_modelcopy}}, and
\code{\link{atlantis_bachchange}}.}
}
\value{
\code{infilename} The name of the parameters file selected.
}
\description{
It looks for the names of the parameters file wanted within the
  batch or shell file used to run Atlantis.
}
\examples{
atlantis_paraselect("C:/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantismain", "-h", "Windows")
atlantis_paraselect("/home/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                    "atlantisNew", "-h", "Linux")

\dontrun{atlantis_paraselect("/Atlantis/AtlantisEEC/AtlantisMSY/COD",
                             "atlantismain", "-h", "Darwin") # for OSX.}
}
